Dijkstra

+++ IntelliJ IDEA project/src/graph/Polaczenie.java
@@ -2,9 +2,9 @@ package graph;
 
 public class Polaczenie {
     public final Wierzcholek cel;
-    public final double odleglosc;
+    public final int odleglosc;
 
-    public Polaczenie(Wierzcholek cel, double odleglosc) {
+    public Polaczenie(Wierzcholek cel, int odleglosc) {
         this.cel = cel;
         this.odleglosc = odleglosc;
     }
